🚀 Show artist color on client
